requirements.txt
iio>=0.0.3
matplotlib>=3.1.2
numpy>=1.18.1
Pillow>=7.0.0
ponomarenko>=0.0.4.post1
rasterio>=1.1.2
scipy>=1.3.0
